# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 隙间（XiJian）本地 API 协议规范
  version: 1.0.0
  description: |
    「隙间」跨平台本地 API 的完整 OpenAPI 3.0 规范。

    ## 架构总览
    隙间本地 API 是 **「OAI 兼容 + 隙间扩展」** 的双协议栈：

    - **OAI 兼容层** (`/v1/*`)：完整实现 OpenAI 官方 API 表面（chat / embeddings / audio / images / video / models / files / fine-tuning 等），第三方客户端（openai-python、langchain、llama-index）可零修改接入。
    - **隙间扩展层** (`/v1/xijian/*`)：以命名空间承载项目特有能力（角色、互动、世界、记忆、保护模块等）。

    ## 传输
    - HTTP + SSE + WebSocket 三栈并行。请求-响应 → HTTP；流式生成 → SSE 或 NDJSON；双向推送（角色主动消息、AI 控制信号）→ WebSocket (`ws://127.0.0.1:{port}/v1/ws`)。
    - 流式协议 SSE 与 NDJSON 同时支持，由客户端通过 `Accept` 头协商。

    ## 鉴权
    - Bearer Token，启动时随机生成，仅 `127.0.0.1` 访问。

    ## 错误格式
    - 双格式共存：`Accept: application/json` 返回 OAI 错误；`Accept: application/json-rpc` 返回 JSON-RPC 2.0 错误。

    ## 参考文档
    - 详细协议见 `docs/api.md`
    - 功能清单见 `docs/Dev. Function List功能清单v2.md`
  license:
    name: 完全开源、无付费、无广告
    identifier: MIT
servers:
- url: http://127.0.0.1:{port}/v1
  description: 本地环回地址（仅监听 127.0.0.1，绝不暴露 0.0.0.0）
  variables:
    port:
      default: '8000'
      description: 由主 UI 进程通过临时文件 /tmp/xijian-<pid>.port 传递
security:
- bearerAuth: []
...
paths:
...
  /xijian/_test/emit:
    post:
      tags:
      - websocket
      operationId: dev_emit
      summary: Dev-only endpoint to publish a fake WS event.
      description: |-
        Dev-only endpoint to publish a fake WS event.

        Guarded by ``XIJIAN_DEV=1`` so it never ships in production.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                data:
                  type: object
              additionalProperties: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: &id017
                type: object
                properties:
                  published:
                    type: string
                  type:
                    type: string
                additionalProperties: true
        '201':
          description: 创建成功
          content:
            application/json:
              schema: *id017
        '403':
          description: 需要 XIJIAN_DEV=1 环境变量
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAIError'
        '400':
          $ref: '#/components/responses/BadRequest'
      x-dev-only: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 启动时随机生成，写入 /tmp/xijian-<pid>.token，仅 127.0.0.1 访问
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OAIError'
...
  schemas:
    OAIError:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
              example: invalid_request_error
            param:
              type: string
            code:
              type: string
              example: invalid_value
      example:
        error:
          message: 'Invalid value for ''temperature'': must be between 0 and 2'
          type: invalid_request_error
          param: temperature
          code: invalid_value
...